graph: Move data::store::IdType to its own file
