+++ graph/src/data/store/mod.rs
@@ -26,6 +26,10 @@ use super::{
     value::Word,
 };
 
+/// Handling of entity ids
+mod id;
+pub use id::IdType;
+
 /// Custom scalars in GraphQL.
 pub mod scalar;
 
@@ -172,33 +176,6 @@ impl ValueType {
     /// Return `true` if `s` is the name of a builtin scalar type
     pub fn is_scalar(s: &str) -> bool {
         Self::from_str(s).is_ok()
-    }
-}
-
-/// The types that can be used for the `id` of an entity
-#[derive(Copy, Clone, Debug, Eq, PartialEq, Hash)]
-pub enum IdType {
-    String,
-    Bytes,
-}
-
-impl<'a> TryFrom<&s::ObjectType> for IdType {
-    type Error = Error;
-
-    fn try_from(obj_type: &s::ObjectType) -> Result<Self, Self::Error> {
-        let base_type = obj_type.field("id").unwrap().field_type.get_base_type();
-
-        match base_type {
-            "ID" | "String" => Ok(IdType::String),
-            "Bytes" => Ok(IdType::Bytes),
-            s => {
-                return Err(anyhow!(
-                    "Entity type {} uses illegal type {} for id column",
-                    obj_type.name,
-                    s
-                ))
-            }
-        }
     }
 }
